pmt: ajouter plusieurs utilisateurs a un projet

+++ pmt/src/main/java/com/pmt/controller/UserController.java
@@ -38,7 +38,7 @@ public class UserController {
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<User> getUser(@PathVariable Integer id) {
+    public ResponseEntity<User> getUser(@PathVariable Long id) {
         try {
             User user = userService.findById(id);
             return ResponseEntity.status(HttpStatus.OK).body(user);
