+++ pmt/src/main/java/com/pmt/service/UserService.java
@@ -6,7 +6,7 @@ import com.pmt.model.User;
 
 public interface UserService {
     List<User> findAll();
-    User findById(Integer id);
+    User findById(Long id);
     User create(User user);
     User login(String email, String password);
 }
